Add files via upload
The objective of the game is to find all the matching pairs of cards.
. The game board consists of 16 cards arranged in a 4x4 grid. 2. Each
card has a front side and a back side. The back side displays a unique letter
(A-H) that forms a pair with another card. Click on any two cards to flip them over and reveal the letters on their back
sides. 

. Only two cards can be flipped at a time.
If two flipped cards do not match, they will be flipped back over after 1
second.
The game ends when all pairs have been successfully matched.